Implemented OpenRC service, updated info command with new metadata, added new goals to README.md
The OpenRC init script was designed to run as the user that owns the executable file.
The SystemD init script was altered to run similarly, instead of copying the executable and running as root.

Init scripts were tested on Fedora and Gentoo Linux respectively.
A minor issue was found with the SystemD script, which should be fixed but I haven't tested it again.
On the Gentoo system I found that iroh-ssh doesn't support OpenDoas.
I wasn't able to test on Windows: cross-compilation is planned.

Signed-off-by: Adam Pimentel <adam.pimentel@protonmail.com>