Test and Fix Windows release, Vendor more of NSSM for transparency, Apply rustfmt
During testing, I found that discovery_n0 is still needed for Windows.

I couldn't find the executable that has the same shasum as the one included in this repo already. For transparency, I will just include the whole unzipped NSSM release directory, which includes references to the version and URL as well as the source code so users can verify themselves that nssm.exe doesn't contain malware.

Signed-off-by: Adam Pimentel <adam.pimentel@protonmail.com>